test(dmm): split Keithley 2400 software tests into per-vendor file
Move the Keithley2400 mocked-transport tests out of test_instro_dmm.py into
tests/dmm/keithley/test_keithley_2400_software.py, mirroring the per-vendor PSU
test layout (and the sibling Agilent split). No test logic changed; imports
trimmed accordingly.

INSTRO-373
